Added basic command logic
